+++ src/main/java/com/tagnumelite/chickens/api/utils/TranslationUtils.java
@@ -1,9 +1,13 @@
 package com.tagnumelite.chickens.api.utils;
 
 import com.tagnumelite.chickens.Chickens;
+import com.tagnumelite.chickens.api.utils.constants.TranslationConstants;
 import net.minecraft.network.chat.Component;
+import net.minecraft.resources.ResourceLocation;
 
-public class TranslateUtils {
+import javax.annotation.Nullable;
+
+public class TranslationUtils {
     public static String GUI(String name) {
         return "gui." + Chickens.MOD_ID + "." + name;
     }
@@ -18,5 +22,10 @@ public class TranslateUtils {
 
     public static Component CGUI_CONTAINER(String name) {
         return Component.translatable(GUI_CONTAINER(name));
+    }
+
+    public static Component CHICKEN_NAME(@Nullable ResourceLocation chickenType) {
+        if (chickenType == null) return Component.translatable(TranslationConstants.NULL_CHICKEN);
+        return Component.translatable("entity." + chickenType.toLanguageKey() + ".name");
     }
 }