+++ src/main/java/com/tagnumelite/chickens/integration/jei/ChickensJeiPlugin.java
@@ -47,7 +47,7 @@ public class ChickensJeiPlugin implements IModPlugin {
 
     @Override
     public void registerRecipeCatalysts(@NotNull IRecipeCatalystRegistration registration) {
-        registration.addRecipeCatalyst(new ItemStack(ModItems.SPAWN_EGG.get()), LayingRecipeCategory.RECIPE_TYPE, BreedingRecipeCategory.RECIPE_TYPE);
+        registration.addRecipeCatalyst(new ItemStack(ModItems.SPAWN_EGG.get()), BreedingRecipeCategory.RECIPE_TYPE, LayingRecipeCategory.RECIPE_TYPE);
     }
 
     @Override
@@ -68,87 +68,4 @@ public class ChickensJeiPlugin implements IModPlugin {
         registration.addRecipes(BreedingRecipeCategory.RECIPE_TYPE, recipeManager.getAllRecipesFor(ModRecipeTypes.BREEDING.get()));
         registration.addRecipes(LayingRecipeCategory.RECIPE_TYPE, Chickens.getChickenManager().getChickens().entrySet().stream().map(ChickenHolder::from).toList());
     }
-
-    /*
-    @Override
-    public void register(IModRegistry registry) {
-        IJeiHelpers jeiHelpers = registry.getJeiHelpers();
-        registry.addRecipeCategories(
-                new LayingRecipeCategory(jeiHelpers.getGuiHelper()),
-                new BreedingRecipeCategory(jeiHelpers.getGuiHelper()),
-                new DropRecipeCategory(jeiHelpers.getGuiHelper()),
-                new ThrowingRecipeCategory(jeiHelpers.getGuiHelper()),
-                new HenhousingRecipeCategory(jeiHelpers.getGuiHelper())
-        );
-        registry.addRecipeHandlers(
-                new LayingRecipeHandler(),
-                new BreedingRecipeHandler(),
-                new DropRecipeHandler(),
-                new ThrowingRecipeHandler(),
-                new HenhousingRecipeHandler()
-        );
-        registry.addRecipes(getLayingRecipes());
-        registry.addRecipes(getBreedingRecipes());
-        registry.addRecipes(getDropRecipes());
-        registry.addRecipes(getThrowRecipes());
-        registry.addRecipes(getHenhouseRecipes());
-    }
-
-    private List<LayingRecipeWrapper> getLayingRecipes() {
-        List<LayingRecipeWrapper> result = new ArrayList<LayingRecipeWrapper>();
-        for (ChickensRegistryItem chicken : ChickensRegistry.getItems()) {
-            result.add(new LayingRecipeWrapper(
-                    new ItemStack(Chickens.spawnEgg, 1, chicken.getId()),
-                    chicken.createLayItem(),
-                    chicken.getMinLayTime(), chicken.getMaxLayTime()
-            ));
-        }
-        return result;
-    }
-
-    private List<DropRecipeWrapper> getDropRecipes() {
-        List<DropRecipeWrapper> result = new ArrayList<DropRecipeWrapper>();
-        for (ChickensRegistryItem chicken : ChickensRegistry.getItems()) {
-            result.add(new DropRecipeWrapper(
-                    new ItemStack(Chickens.spawnEgg, 1, chicken.getId()),
-                    chicken.createDropItem()
-            ));
-        }
-        return result;
-    }
-
-    private List<BreedingRecipeWrapper> getBreedingRecipes() {
-        List<BreedingRecipeWrapper> result = new ArrayList<BreedingRecipeWrapper>();
-        for (ChickensRegistryItem chicken : ChickensRegistry.getItems()) {
-            if (chicken.isBreedable()) {
-                //noinspection ConstantConditions
-                result.add(new BreedingRecipeWrapper(
-                        new ItemStack(Chickens.spawnEgg, 1, chicken.getParent1().getId()),
-                        new ItemStack(Chickens.spawnEgg, 1, chicken.getParent2().getId()),
-                        new ItemStack(Chickens.spawnEgg, 1, chicken.getId()),
-                        ChickensRegistry.getChildChance(chicken)
-                ));
-            }
-        }
-        return result;
-    }
-
-    private List<ThrowingRecipeWrapper> getThrowRecipes() {
-        List<ThrowingRecipeWrapper> result = new ArrayList<ThrowingRecipeWrapper>();
-        for (ChickensRegistryItem chicken : ChickensRegistry.getItems()) {
-            if (chicken.isDye()) {
-                result.add(new ThrowingRecipeWrapper(
-                        new ItemStack(Chickens.coloredEgg, 1, chicken.getDyeMetadata()),
-                        new ItemStack(Chickens.spawnEgg, 1, chicken.getId())));
-            }
-        }
-        return result;
-    }
-
-    private List<HenhousingRecipeWrapper> getHenhouseRecipes() {
-        List<HenhousingRecipeWrapper> henhouseRecipes = new ArrayList<HenhousingRecipeWrapper>();
-        henhouseRecipes.add(new HenhousingRecipeWrapper(new ItemStack(Blocks.HAY_BLOCK), new ItemStack(Blocks.DIRT)));
-        return henhouseRecipes;
-    }
-     */
 }